Carouselのcomponent化＆Destination,Tour BlogのCarousel化
